Refactor : jwt
