+++ src/middlewares/jwt/jwtAuthMiddleware.js
@@ -1,5 +1,5 @@
 const jwt = require("jsonwebtoken");
-const { SECRET_KEY } = require("../../config/config");
+const { SECRET_KEY } = require("../../config");
 
 function authenticateToken(req, res, next) {
   const authHeader = req.headers.authorization;
